main_service/p_project/requirements.txt
asgiref==3.8.1
build==1.2.2.post1
click==8.1.7
distlib==0.3.9
Django==5.1.4
djangorestframework==3.15.2
filelock==3.16.1
packaging==24.2
pip-tools==7.4.1
platformdirs==4.3.6
pyproject_hooks==1.2.0
setuptools==75.6.0
sqlparse==0.5.3
virtualenv==20.27.0
wheel==0.45.1
